+++ Ircu.c
@@ -362,7 +362,7 @@ int
 snick_cmd (const char *oldnick, const char *newnick)
 {
 	Change_User (finduser (oldnick), newnick);
-	sts (":%s %s %s %d", oldnick, MSG_NICK, newnick, me.now);
+	sts (":%s %s %s %d", oldnick, MSG_NICK, newnick, (int) me.now);
 	return 1;
 }
 
@@ -429,7 +429,7 @@ ssvshost_cmd (const char *who, const char *vhost)
 int
 ssvinfo_cmd ()
 {
-	sts ("SVINFO 5 3 0 :%d", me.now);
+	sts ("SVINFO 5 3 0 :%d", (int) me.now);
 	return 1;
 }
 
@@ -460,7 +460,7 @@ sakill_cmd (const char *host, const char *ident, const char *setby, const int le
 	hash_scan_begin (&ss, sh);
 	while ((sn = hash_scan_next (&ss)) != NULL) {
 		s = hnode_get (sn);
-		sts (":%s %s %s %lu %s %s :%s", setby, MSG_KLINE, s->name, length, ident, host, ircd_buf);
+		sts (":%s %s %s %lu %s %s :%s", setby, MSG_KLINE, s->name, (unsigned long)length, ident, host, ircd_buf);
 	}
 	return 1;
 }
