+++ hybrid7.c
@@ -362,7 +362,7 @@ int
 snick_cmd (const char *oldnick, const char *newnick)
 {
 	Change_User (finduser (oldnick), newnick);
-	sts (":%s %s %s %d", oldnick, MSG_NICK, newnick, me.now);
+	sts (":%s %s %s %d", oldnick, MSG_NICK, newnick, (int)me.now);
 	return 1;
 }
 
@@ -434,7 +434,7 @@ sinvite_cmd (const char *from, const char *to, const char *chan) {
 int
 ssvinfo_cmd ()
 {
-	sts ("SVINFO 5 3 0 :%d", me.now);
+	sts ("SVINFO 5 3 0 :%d", (int)me.now);
 	return 1;
 }
 
@@ -454,7 +454,7 @@ sakill_cmd (const char *host, const char *ident, const char *setby, const int le
 	va_start (ap, reason);
 	ircvsnprintf (ircd_buf, BUFSIZE, reason, ap);
 	va_end (ap);
-	sts (":%s %s * %lu %s %s :%s", setby, MSG_KLINE, length, ident, host, ircd_buf);
+	sts (":%s %s * %lu %s %s :%s", setby, MSG_KLINE, (unsigned long)length, ident, host, ircd_buf);
 	return 1;
 }
